refactor: consolidate files / utils
- remove pseudo duplicated Files by merging into FilesClient and / or fs-utils

closes #375
